requirements.txt
numpy==1.13.3
pandas==0.21.0
scipy==0.19.1
scikit-learn==0.19.1
python-forecastio==1.3.5